requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
config==0.5.1
contourpy==1.3.2
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dpcpp-cpp-rt==2025.1.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.2
humanfriendly==10.0
idna==3.10
ImageHash==4.3.2
imagesize==1.4.1
intel-cmplr-lib-rt==2025.1.1
intel-cmplr-lib-ur==2025.1.1
intel-cmplr-lic-rt==2025.1.1
intel-opencl-rt==2025.1.1
intel-openmp==2025.1.1
intel-pti==0.12.3
intel-sycl-rt==2025.1.1
intel_extension_for_pytorch==2.8.10+xpu
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lxml==6.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
mkl==2025.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.3
numpy==2.3.1
onemkl-sycl-blas==2025.1.0
onemkl-sycl-dft==2025.1.0
onemkl-sycl-lapack==2025.1.0
onemkl-sycl-rng==2025.1.0
onemkl-sycl-sparse==2025.1.0
onnxruntime==1.22.1
openai==1.97.1
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.1
packaging==25.0
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.17.0
pillow==11.0.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymilvus==2.6.0
pymilvus.model==0.3.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-triton-xpu==3.4.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.1
setuptools==70.2.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.41
sympy==1.13.3
tbb==2022.1.0
tcmlib==1.3.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.8.0+xpu
torchvision==0.23.0+xpu
tqdm==4.67.1
transformers==4.54.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
umf==0.10.0
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
